chore: seperated native and web views for index
